+++ src/main/java/kr/sparta/rchive/domain/backoffice/respoonse/BackofficeResponseCode.java
@@ -1,5 +1,7 @@
-package kr.sparta.rchive.domain.user.response;
+package kr.sparta.rchive.domain.backoffice.respoonse;
 
+
+import kr.sparta.rchive.domain.user.response.UserResponseCode;
 import kr.sparta.rchive.global.response.ResponseCode;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,9 +9,10 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @AllArgsConstructor
-public enum ProfileResponseCode implements ResponseCode {
+public enum BackofficeResponseCode implements ResponseCode {
     // OK 200
-
+    OK_GET_LAST_SELECT_ROLE(HttpStatus.OK, "BACKOFFICE-008", "백오피스: 마지막으로 선택한 권한 조회 성공"),
+    OK_GET_PROFILE(HttpStatus.OK, "BACKOFFICE-009","백오피스: 프로필 조회 성공")
     ;
 
     private final HttpStatus httpStatus;
